EDM-168 Parametrizable templates
