+++ libs/i18n/locales/en/translation.json
@@ -26,6 +26,7 @@
   "Repository Details": "Repository Details",
   "Resource sync": "Resource sync",
   "No devices": "No devices",
+  "Copy text": "Copy text",
   "New label": "New label",
   "Add label": "Add label",
   "Unexpected error occurred": "Unexpected error occurred",
@@ -39,7 +40,6 @@
   "Discard changes": "Discard changes",
   "Stay here": "Stay here",
   "Are you sure you want to leave the current page? Unsaved changes will be lost.": "Are you sure you want to leave the current page? Unsaved changes will be lost.",
-  "Copy text": "Copy text",
   "Fleet not found": "Fleet not found",
   "We could not find the fleet with id <1>{id}</1>": "We could not find the fleet with id <1>{id}</1>",
   "Device not found": "Device not found",
@@ -179,6 +179,8 @@
   "Repository": "Repository",
   "Select a repository": "Select a repository",
   "Create repository": "Create repository",
+  "Using template variables": "Using template variables",
+  "Add a variable using <1>{templateCode}</1> and it will be applied based on the devices labels. Template variables can only be used in Inline and Git configurations.": "Add a variable using <1>{templateCode}</1> and it will be applied based on the devices labels. Template variables can only be used in Inline and Git configurations.",
   "The target system image for this fleet's devices.": "The target system image for this fleet's devices.",
   "Must be either an OCI image ref (e.g. \"quay.io/redhat/rhde:9.3\") or ostree ref (e.g. \"https://ostree.fedoraproject.org/iot?ref=fedora/stable/x86_64/iot\"). Keep this empty if you do not want to manage your OS from fleet.": "Must be either an OCI image ref (e.g. \"quay.io/redhat/rhde:9.3\") or ostree ref (e.g. \"https://ostree.fedoraproject.org/iot?ref=fedora/stable/x86_64/iot\"). Keep this empty if you do not want to manage your OS from fleet.",
   "Device labels": "Device labels",
@@ -206,7 +208,7 @@
   "Create fleet": "Create fleet",
   "Failed to determine the number of selected devices": "Failed to determine the number of selected devices",
   "Try again": "Try again",
-  "{{ count }} devices matching the labels were selected._one": "{{ count }} devices matching the labels was selected.",
+  "{{ count }} devices matching the labels were selected._one": "{{ count }} device matching the labels was selected.",
   "{{ count }} devices matching the labels were selected._other": "{{ count }} devices matching the labels were selected.",
   "Add labels to select devices to be included in this fleet.": "Add labels to select devices to be included in this fleet.",
   "Fleet labels": "Fleet labels",
