+++ libs/ui-components/src/components/common/CopyButton.tsx
@@ -1,0 +1,30 @@
+import * as React from 'react';
+import { CopyIcon } from '@patternfly/react-icons/dist/js/icons/copy-icon';
+import { Button, ButtonProps, Icon, Tooltip } from '@patternfly/react-core';
+
+import { useTranslation } from '../../hooks/useTranslation';
+
+const CopyButton = ({ text, variant }: { text: string; variant?: ButtonProps['variant'] }) => {
+  const { t } = useTranslation();
+
+  const onCopy = () => {
+    void navigator.clipboard.writeText(text);
+  };
+
+  return (
+    <Tooltip content={t('Copy text')}>
+      <Button
+        variant={variant || 'plain'}
+        isInline={variant === 'link'}
+        icon={
+          <Icon size="sm">
+            <CopyIcon onClick={onCopy} />
+          </Icon>
+        }
+        aria-label={t('Copy text')}
+      />
+    </Tooltip>
+  );
+};
+
+export default CopyButton;
