+++ libs/ui-components/src/components/common/ResourceLink.tsx
@@ -1,9 +1,7 @@
 import * as React from 'react';
-import { CopyIcon } from '@patternfly/react-icons/dist/js/icons/copy-icon';
-import { Button, Icon, Tooltip } from '@patternfly/react-core';
 
-import { useTranslation } from '../../hooks/useTranslation';
 import { Link, type RouteWithPostfix } from '../../hooks/useNavigate';
+import CopyButton from './CopyButton';
 
 import './ResourceLink.css';
 
@@ -26,29 +24,6 @@ export const getDisplayText = (name: string | undefined) => {
   return `${name.substring(0, 6)}...${name.substring(name.length - 7)}`;
 };
 
-const CopyNameButton = ({ name }: { name: string }) => {
-  const { t } = useTranslation();
-
-  const onCopy = () => {
-    void navigator.clipboard.writeText(name);
-  };
-
-  return (
-    <Tooltip content={t('Copy text')}>
-      <Button
-        variant="plain"
-        isInline
-        icon={
-          <Icon size="sm">
-            <CopyIcon onClick={onCopy} />
-          </Icon>
-        }
-        aria-label={t('Copy text')}
-      />
-    </Tooltip>
-  );
-};
-
 const ResourceLink = ({ id, name, variant = 'shortened', routeLink }: ResourceDisplayLinkProps) => {
   const nameOrId = name || id;
   const displayText = getDisplayText(nameOrId);
@@ -59,7 +34,7 @@ const ResourceLink = ({ id, name, variant = 'shortened', routeLink }: ResourceDi
   return (
     <span className={`fctl-resource-link fctl-resource-link__${variant}`}>
       {routeLink ? <Link to={{ route: routeLink, postfix: id }}>{textEl}</Link> : <>{textEl}</>}
-      {showCopy && nameOrId && <CopyNameButton name={nameOrId} />}
+      {showCopy && nameOrId && <CopyButton text={nameOrId} />}
     </span>
   );
 };
